// prisma/types.ts
export type MediaType = 'Video' | 'Image';

export type UserType = 'Student' | 'Teacher' | 'Admin';

export type QuestionType =
  | 'button'
  | 'checkbox'
  | 'color'
  | 'date'
  | 'datetime_local'
  | 'email'
  | 'file'
  | 'hidden'
  | 'image'
  | 'month'
  | 'number'
  | 'password'
  | 'radio'
  | 'range'
  | 'reset'
  | 'search'
  | 'submit'
  | 'tel'
  | 'text'
  | 'time'
  | 'url'
  | 'week';

export type LocationType = 'Activities' | 'Modules' | 'Quizzes';